Add Cursor class in Core IO module
A new file Cursor.hpp has been added to the IO sub-directory in the Core module. It's included in the CMakeLists for the Core module to ensure its compilation. The Cursor class provides a way to safely navigate over index-able containers, managing and securing reading operations. This is required to enhance control and prevent potential over-reading errors in our IO operations.