Refactor AsVector method to use templates
Updated the AsVector method in DynamicByteBuffer class to use templates. This change allows returning vector of different data types, not just bytes. It can be more convenient for handling different data in the buffer. Previously, this method could only return vector of bytes, that was not flexible enough for all usage scenarios.