Add SizedContainer concept to Core Concepts
This commit introduces a new "SizedContainer" concept to the Core Concepts in our codebase. This concept verifies if a given class has a "Size" method returning size_t. This will consistently enforce container size format rules throughout our code. By introducing this concept, it will modularize the code and make it easier to add or modify types that require size.