"Added Write function to Cursor class in IO module"
This commit adds two versions of a Write function to the Cursor class in the IO module. The first version simply writes bytes to the container at the current cursor position. The second version is similar but includes an assertion to check whether the operation would exceed the size of the container when SizedContainer is present or the container type is vector. These new functions enhance the functionality of the Cursor class by providing a way to write data to its underlying container.