Refactor template in Cursor class
Changed the template from ReferenceIndexable to Indexable in the Cursor class within the IO namespace of Strawberry Core. This change was necessary because we found that not all insertions to the Cursor class need to satisfy ReferenceIndexable's requirement of being a reference type. This opens up more flexibility for future container types that may interact with the Cursor class.