Add IndexedType concept to Concepts.hpp
The new template, IndexedType, has been added to Concepts.hpp. This template works with Indexable types and can be helpful in determining the type of elements at a specific index within Indexable objects. This can further enhance readability and maintainability in situations where we need to deal with Indexable types.