Rename 'Circular' class to 'CircularAdapter'
The class 'Circular' was renamed to 'CircularAdapter' and all references to it were also updated. This naming more accurately reflects the purpose of the class, which is to act as an adapter for circular data structures. The constructor for the empty container was also removed as it was deemed redundant.