Refactor Clock class to start on object instantiation
The Clock class constructor has been updated to start the clock if the flag is set. Previously, the argument name 'run' was misleading and didn't automatically run the clock as expected. It is now aptly renamed 'start' and causes the clock to start immediately upon Clock object creation if set to true."