Add constructor for DynamicPeriodic in Periodic.hpp
A new constructor has been added to the DynamicPeriodic class within the Periodic.hpp file. This constructor accepts a single parameter 'value', which it then caps to 'max', included as a part of this class. This makes it easier to establish a new DynamicPeriodic object without explicitly stating the maximum possible value.