Remove redundant RTP/Header.cpp file from project
The RTP/Header.cpp file was found to be redundant and not serving any purpose within the project, hence it was removed. This step is part of the ongoing effort to clean up the code and improve its maintainability. Also, related reference in the CMakeLists.txt file is deleted to keep the build configuration up to date.