Add ScopedTimer to StrawberryCore
A ScopedTimer class has been added to the StrawberryCore, it includes a corresponding .cpp file. This new feature will allow precise time measurements for scope-duration, useful for performance profiling and benchmarking. It has also been added to the CMakeLists.txt, ensuring it's properly built along with the rest of the Core.