Add NullOpt to Option class for better usability
Added a new class NullOpt_t and an instance NullOpt to the Option functionality in Strawberry's Core system. This newly incorporated feature allows for better handling of "null" situations and enhances the usability of the Option class. Now the "null" instances can be represented by using NullOpt, removing the need for exception handling for "null" cases. Also updated the '=' operator to reset the Option when assigned a NullOpt.