Add 'AndThen' method to 'Option' class
The 'AndThen' method was added to the 'Option' class to allow for more streamlined chaining of operations on the Option object. The method receives a functor and invokes it if the Option has value. The result of this invocation, which is expected to be of the 'Option' type, is returned. If the initial Option has no value or the functor invocation does not return an Option with value, 'NullOpt' is returned. This enhancement enriches the API of the 'Option' class and facilitates more expressive use-cases for this class.