Refactor Metronome class to use preemption instead of eagerness
In this commit, the `Metronome` class has been refactored to use a concept of "preemption" instead of "eagerness". The method `SetEagerness()` has been replaced with a new method `SetPreemption()`. The variable `mEagerness` has also been changed into `mPreemption`. This change was made to give a clearer understanding of the functionality. Furthermore, related calculations in the code have been adapted accordingly to comply with this change.