Add Variant helper class to Core Utils
The Variant class is added to Core_Utils providing a std::variant-like interface with additional utility functions. These functions allow users to directly check if the Variant contains a value, if the value is of a specific type, and to get the value with automatic type checking. This change should make working with "optional" values safer and easier throughout the project. Also, the new file Variant.hpp is now included in the CMakeLists.txt for StrawberryCore.