Add Lazy class for delayed initialization
Added a new file Lazy.hpp which is a utility class in the Core library. It implements the 'Lazy' class template for deferred initialization of objects. This resolution strategy is used when the cost of initialization is high and resources need to be effectively utilized by only initializing when it's necessary. This class forces an object to hold its initialization until a 'Get' method is called.