Add Copy-On-Write utility class
A new utility class `COW` (Copy-On-Write) has been added to the Strawberry Core library. This class uses shared pointers and templates to create an efficient implementation of the Copy-On-Write pattern. This allows the class to be both memory and performance efficient, as deep copies are only made when necessary. This class will be extensively used in the library to reduce unnecessary object copies.